Validar envio de archivos

+++ app/Http/Controllers/API/SeccionesController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\secciones\SeccionesRequest;
 use App\models\Seccion;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use App\Helpers\StorageFiles;
+use App\Http\Requests\secciones\SeccionesUpdateRequest;
 
 class SeccionesController extends Controller
 {
@@ -17,7 +19,7 @@ class SeccionesController extends Controller
      */
     public function index()
     {
-        $secciones = Seccion::all();
+        $secciones = Seccion::paginate(5);
 
         return response()->json(['secciones' => $secciones], 200);
     }
@@ -33,10 +35,13 @@ class SeccionesController extends Controller
     public function store(SeccionesRequest $request)
     {
         DB::beginTransaction();
-
-        if (!$request->has('errors')) {
+        if (!$request->has('errors') && $request->hasFile('imagen')) {
             try {
-                Seccion::create($request->validated());
+                $validated = $request->validated();
+                $storage_files = new StorageFiles('secciones', $validated['nombre'], $request->file('imagen'));
+                $path = $storage_files->createFile();
+                $body = $this->getBodySecciones($validated['nombre'], $path);
+                Seccion::create($body);
                 DB::commit();
                 return response()->json('¡Sección guardada exitosamente!', 200);
             } catch (\Exception $e) {
@@ -57,9 +62,12 @@ class SeccionesController extends Controller
      */
     public function show($id)
     {
-        $sensor = Seccion::findOrFail($id);
-
-        return response()->json($sensor, 200);
+        try {
+            $seccion = Seccion::findOrFail($id);
+            return response()->json($seccion, 200);
+        } catch (\Exception $e) {
+            return response()->json(['message' => $e->getMessage()], 500);
+        }  
     }
 
     /**
@@ -70,7 +78,7 @@ class SeccionesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(SeccionesRequest $request, $id)
+    public function update(SeccionesUpdateRequest $request, $id)
     {
         if(!$request->has('errors')) 
         {
@@ -78,6 +86,11 @@ class SeccionesController extends Controller
 
             try {
                 $seccion = Seccion::find($id);
+                $validated = $request->validated();
+                $file = $request->hasFile('imagen') ?  $request->file('imagen') : null;
+                $name = $validated['nombre'] !== $seccion-> nombre ? $validated['nombre'] : null;
+                $storage_files = new StorageFiles('secciones', $name,  $file);
+                $path = $storage_files->createFile();
                 $seccion->update($request->validated());
                 DB::commit();
                 return response()->json(['seccion' => $seccion, 'message' => '¡Sección actualizada exitosamente!'], 200);
@@ -102,12 +115,32 @@ class SeccionesController extends Controller
 
         try {
             $seccion = Seccion::findOrFail($id);
-            $seccion->delete();
-            DB::commit();
-            return response()->json('¡Sección eliminada exitosamente!', 200);
+            $storage_files = new StorageFiles($seccion->imagen);
+            $path = $storage_files->deleteFile();
+            if($path) {
+                $seccion->delete();
+                DB::commit();
+                return response()->json('¡Sección eliminada exitosamente!', 200);
+            } else {
+                return response()->json('¡No se ha encontrado el archivo!', 404);
+            }
         } catch (\Exception $e) {
             DB::rollback();
             return response(['message' => $e->getMessage()]);
         }
     }
+
+    /**
+     * Funcion para armarel cuerpo de la peticion de la seccion
+     * al ser un archivo se le enviara la ruta y el nombre del mismo
+     */
+    private function getBodySecciones($name, $path)
+    {
+        $body = [
+            'nombre' => $name,
+            'imagen' => $path
+        ];
+
+        return $body;
+    } 
 }
